Fix SDDM template entry header, sddm-process rm fix, add existence check to tmp-clear

+++ script/helpers/tmp-clear.sh
@@ -6,6 +6,10 @@ if [[ -z "$FEATHERT" ]]; then
 	echo "FEATHERT is not set!"
 	exit 1
 fi
+if ! [[ -d "$FEATHERT" ]]; then
+	echo "Temp is not spawned!"
+	exit 0
+fi
 if [[ "--if-empty" != "$1" ]] || ! [ "$(ls -A "$FEATHERT")" ]; then
 	rm -rf "$FEATHERT"
 fi
